// N/client/src/module/Admin/admin.validator.tsx
export interface CouponProps {
    _id: string;
    code: string;
    discountType: 'percentage' | 'fixed'; // assuming you may have fixed amount discount as well
    discountValue: number;
    validFrom: string; // or Date if you plan to convert it
    validUntil: string; // or Date
    isActive: boolean;
    applicableCategories: string[];
    __v: number;
}
// ...
export interface CouponResponse {
    code: number;
    data: CouponProps[];
    message: string;
    options: {
        total: number;
    };
    status: string;
}

export interface UserResponse {
    code: number;
    data: listUserProps[];
    message: string;
    options: {
        total: number;
    };
    status: string;
}

export interface CategoryResponse {
    code: number;
    data: ListCategoryDetails[];
    message: string;
    options: {
        total: number;
    };
    status: string;
}